requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.3.0
bech32==1.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cosmpy==0.11.1
distlib==0.4.0
ecdsa==0.19.1
filelock==3.19.1
frozenlist==1.7.0
googleapis-common-protos==1.70.0
grpcio==1.74.0
h11==0.16.0
idna==3.10
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
multidict==6.6.4
platformdirs==4.4.0
propcache==0.3.2
protobuf==5.29.5
pycryptodome==3.23.0
pydantic==2.11.9
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
sortedcontainers==2.4.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uagents==0.22.8
uagents-core==0.3.8
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.34.0
yarl==1.20.1

python-dotenv
openai
ddgs